requirements.txt
streamlit
google-generativeai
python-dotenv
sentence-transformers
torch
scikit-learn
transformers
